Add a change detection mechanism to rospy
If the period or transportation delay on a topic/connection
changes, this can be an indication of a malfunction.

This patch adds a change detector to rospy that observers those
two metrics and sends out a change event if they change drastically.

This patch depends on the the TopicStatistics enhancement for rospy.